requirements.txt
# Core libraries
numpy==1.24.3
opencv-python==4.8.0.76
face-recognition==1.3.0
dlib==19.24.2
scipy==1.10.1

# Web framework and utilities
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
marshmallow==3.20.1

# Database
SQLAlchemy==2.0.21
alembic==1.12.0

# Security
PyJWT==2.8.0
python-dotenv==1.0.0

# Utilities
Pillow==10.0.1
pandas==2.1.1
matplotlib==3.8.0